Require cool_off_time to be a whole number of seconds
Strings, nil, zero and negatives were accepted at registration and failed
later: a TypeError from the data store on the transition to red, or a
light that went straight back into recovery. Sub-second values made the
light probe on nearly every call instead of breaking the circuit, and a
fractional part above that is marginal against the window failures are
measured over.

Reject anything but an Integer >= 1 when the config is applied, and make
the default 60 rather than 60.0.

+++ features/stoplight/support/configure_light_world.rb
@@ -54,7 +54,7 @@ module ConfigureLightWorld
   end
 
   def configure_cool_off_time(value, settings)
-    settings[:cool_off_time] = value.to_f
+    settings[:cool_off_time] = value.to_i
   end
 
   def configure_threshold(value, settings)
